refactor(config): extract validation to separate file and enhance structure
